refactor: Split up PostsController into multiple files
