+++ app/Http/Controllers/Posts/CountByMonth.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Http\Controllers\Posts;
+
+trait CountByMonth
+{
+}
