+++ app/Http/Controllers/Posts/ListPosts.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers\Posts;
+
+use Illuminate\Http\Request;
+
+trait ListPosts
+{
+    public function listPosts(Request $request)
+    {
+        $posts = self::_resolvedPosts();
+
+        return self::_after($request, $posts, null, false);
+    }
+}
